+++ subprojects/bounded_context/accessibility/src/output-adapter/kotlin/club/staircrusher/accessibility/output_adapter/service/InMemoryPlaceService.kt
@@ -2,13 +2,12 @@ package club.staircrusher.accessibility.output_adapter.service
 
 import club.staircrusher.accessibility.domain.model.Place
 import club.staircrusher.accessibility.domain.service.PlaceService
-import club.staircrusher.place.application.PlaceApplicationService
 
 class InMemoryPlaceService(
-    private val placeApplicationService: PlaceApplicationService,
+    private val placeService: club.staircrusher.place.application.service.PlaceService,
 ) : PlaceService {
     override fun findPlace(placeId: String): Place? {
-        return placeApplicationService.findPlace(placeId)?.let { Place(
+        return placeService.findPlace(placeId)?.let { Place(
             id = it.id,
             buildingId = it.building.id,
         ) }
