+++ subprojects/bounded_context/place/src/domain/kotlin/club/staircrusher/place/domain/model/BuildingAddress.kt
@@ -1,4 +1,4 @@
-package club.staircrusher.place.domain.entity
+package club.staircrusher.place.domain.model
 
 data class BuildingAddress(
     val siDo: String,
