+++ subprojects/bounded_context/place_search/src/application/kotlin/club/staircrusher/place_search/application/service/PlaceSearchService.kt
@@ -1,12 +1,12 @@
-package club.staircrusher.place_search.application
+package club.staircrusher.place_search.application.service
 
 import club.staircrusher.place_search.domain.model.Place
-import club.staircrusher.place_search.domain.service.AccessibilityService
-import club.staircrusher.place_search.domain.service.PlaceService
+import club.staircrusher.place_search.application.port.out.AccessibilityService
+import club.staircrusher.place_search.application.port.out.PlaceService
 import club.staircrusher.stdlib.geography.Length
 import club.staircrusher.stdlib.geography.Location
 
-class PlaceSearchApplicationService(
+class PlaceSearchService(
     private val placeService: PlaceService,
     private val accessibilityService: AccessibilityService,
 ) {
